refactor(lsp): split LSPManager into focused files (client registry, request router, indexing orchestrator, command resolver); preserve behavior and tests
